refactor submit workflows into focused files
